- advanced coordinate control for overlays over DTA_Fullscreen images.
Manually merged, there were severe conflicts.

+++ src/textures/texture.cpp
@@ -1621,7 +1621,7 @@ float FTexCoordInfo::TextureAdjustWidth() const
 	{
 		float tscale = fabs(mTempScale.X);
 		if (tscale == 1.f) return (float)mRenderWidth;
-		else return mWidth / fabs(tscale);
+		else return mWidth / fabsf(tscale);
 	}
 	else return (float)mWidth;
 }
